Improve error messages
Currently when an exception is raised, we will only see “an error
occurred” message. This is not very useful since it provides no
instructions on what to do.

Instead we should point user to check the logs or even point out
that the logs are read-only when that is the case.