controllers: Inject Authentication helper
One more step in decoupling selfoss classes.

We mark helpers\Authenticate class shared since we only want a single
instance of it – mainly because the helper configures session cookie
before starting a session, and the configuration cannot be repeated
after the session has been started.

We still need to let the DI container construct the database classes
so that we can inject Auth helper there too. Until then, we will
keep registering it with the service locator.